changed daysleft to int8

+++ templateinfo/struct.go
@@ -3,7 +3,7 @@ package templateinfo
 type TemplateInfo struct {
 	IsToday   bool
 	Name      string
-	DaysUntil int32
+	DaysUntil int8
 	Date      string
 	GifURL    string
 	Day       int
@@ -12,7 +12,7 @@ type TemplateInfo struct {
 	Weekday   string
 }
 
-func NewTemplateInfo(name string, is_today bool, days_until int32, date string, gif_url string, day int, month string, year int, week_day string) TemplateInfo {
+func NewTemplateInfo(name string, is_today bool, days_until int8, date string, gif_url string, day int, month string, year int, week_day string) TemplateInfo {
 	var template_info TemplateInfo
 	template_info.Name = name
 	template_info.IsToday = is_today
